feat: move `resolve` option to `url` option
Closes #914.
