+++ test/fixtures/url-as-string/img.png
