+++ test/fixtures/url-as-string/other-query.css
