+++ test/fixtures/url-as-string/url.css
